restored untracking functionality
